backend/requirements.txt
# FastAPI and dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-decouple==3.8

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Validation and serialization
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# Authentication
python-jose==3.3.0
bcrypt==4.1.2

# HTTP client for web scraping
httpx==0.25.2
beautifulsoup4==4.12.2
lxml==4.9.3
selenium==4.15.2
fake-useragent==1.4.0
requests-ratelimiter==0.4.2
retry==0.9.2

# Background tasks
celery==5.3.4
redis==5.0.1

# CORS
fastapi-cors==0.0.6

# File uploads
aiofiles==23.2.1

# Logging
structlog==23.2.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Development
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Monitoring
sentry-sdk[fastapi]==1.38.0

# Payment processing (Stripe)
stripe==7.8.0

# Google Maps
googlemaps==4.10.0

# Email
fastapi-mail==1.4.1

# Image processing
Pillow==10.1.0

# Date/time utilities
python-dateutil==2.8.2

# Environment configuration
python-dotenv==1.0.0